requirements.txt
beautifulsoup4>=4.12.0
supabase>=2.4.0
python-dotenv>=1.0.1
openai
pinecone